+++ Scroller/Scroller.js
@@ -57,26 +57,28 @@ let Scroller = (props) => {
 
 		resizeContextProps,
 		scrollContainerProps,
-		scrollInnerContainerProps,
 		scrollContentWrapperProps,
 		scrollContentProps,
+		ScrollToTopButton,
 		verticalScrollbarProps,
 		horizontalScrollbarProps
 	} = useScroll(props);
 
-	const {focusableBodyProps, themeScrollContentProps} = useThemeScroller(props, scrollContentProps);
+	const {
+		focusableBodyProps,
+		themeScrollContentProps
+	} = useThemeScroller(props, scrollContentProps);
 
 	// Render
 	const scrollContainer = (
 		<ResizeContext.Provider {...resizeContextProps}>
 			<div {...scrollContainerProps}>
-				<div {...scrollInnerContainerProps}>
-					<ScrollContentWrapper {...scrollContentWrapperProps}>
-						<UiScrollerBasic {...themeScrollContentProps} ref={scrollContentHandle} />
-					</ScrollContentWrapper>
-				</div>
+				<ScrollContentWrapper {...scrollContentWrapperProps}>
+					<UiScrollerBasic {...themeScrollContentProps} ref={scrollContentHandle} />
+				</ScrollContentWrapper>
 				{isVerticalScrollbarVisible ? <Scrollbar {...verticalScrollbarProps} /> : null}
 				{isHorizontalScrollbarVisible ? <Scrollbar {...horizontalScrollbarProps} /> : null}
+				<ScrollToTopButton />
 			</div>
 		</ResizeContext.Provider>
 	);