+++ CHANGELOG.md
@@ -2,6 +2,52 @@
 
 The following is a curated list of changes in the Enact sandstone module, newest changes on the top.
 
+## [unreleased]
+
+### Added
+
+- `sandstone/Panels` and `sandstone/Panels.WizardPanel` props `onTransition` and `onWillTransition`
+- `sandstone/Panels.WizardPanel` prop `noAnimation` to suppress view transition animation
+- `sandstone/Slider` and `sandstone/ProgressBar` prop `showAnchor` to display anchor based on `progressAnchor` value
+
+## [1.0.0-alpha.9] - 2020-04-20
+
+### Deprecated
+
+- `sandstone/TabLayout` prop `tabs`, to be removed in beta.1. Use `sandstone/TabLayout.Tab` instead.
+
+### Added
+
+- `sandstone/Panels.Panel` prop `featureContent` to minimize the panel visuals to feature the content more prominently
+- `sandstone/TabLayout.Tab` for configuring `TabLayout` tab contents
+
+### Fixed
+
+- `sandstone/Button` styles for `selected`
+- `sandstone/Switch` sizing and positioning in large text mode
+- `sandstone/Checkbox` and `sandstone/RadioItem` styling when disabled and focused
+
+## [1.0.0-alpha.8] - 2020-04-14
+
+### Deprecated
+
+- `sandstone/Panels.Header` props `headerInput` and `showInput`, to be removed in 1.0.0-beta.1
+
+### Added
+
+- `sandstone/Panels.FlexiblePopupPanels` for a flexible size pop-up Panels experience
+- `sandstone/Panels` and `sandstone/Panels.Header` props `backButtonAriaLabel`, `backButtonBackgroundOpacity`, `closeButtonAriaLabel`, `closeButtonBackgroundOpacity`, `noBackButton`, `noCloseButton`, `onBack`, and `onClose`
+
+### Changed
+
+- `sandstone/Panels.OptionPanels` to `sandstone/Panels.FixedPopupPanels`
+- `sandstone/Scroller` and `sandstone/VirtualList` overscroll effect to bounce
+- `sandstone/Picker` horizontal joined behavior and style for updated GUI
+
+### Fixed
+
+- `sandstone/TabLayout` to not select a previously focused tab after switching from 5-way to pointer mode
+
 ## [1.0.0-alpha.7] - 2020-04-06
 
 ### Added
@@ -10,7 +56,7 @@ The following is a curated list of changes in the Enact sandstone module, newest
 
 ### Changed
 
-- `sandstone/Picker`, `sandstone/ProgressBar.ProgressBarTooltip`, and `standstone/Steps` to use a number font for numeric content
+- `sandstone/Picker`, `sandstone/ProgressBar.ProgressBarTooltip`, and `sandstone/Steps` to use a number font for numeric content
 
 ### Fixed